cambios en el menu para cada usuario y los carteles de aviso de sesion inicia y sesion salida.

+++ js/validarLogin.js
@@ -8,22 +8,24 @@ function EnviarDatos() {
 
     funcionAjax.then(function(dato) {
         if (dato.status == "200" && dato.tipo == "Cliente") {
-            swal(
-                'USUARIO VÁLIDO!',
-                'Usted está registrado en la base de datos!',
-                'success', { button: 'aceptar', }
-            ).then(function() {
+            swal({
+               title: 'USUARIO VÁLIDO',
+               text: 'Usted está registrado/a',
+               timer: 1000,
+               showConfirmButton: false                   
+            }).then(function() {
                 localStorage.setItem("cliente", dato.Usuario);
                 window.location.replace("../enlaces/restauranteCliente.html");
             }, function() {
                 swal('Algo inesperado ocurrio');
             });
         } else if (dato.status == "200" && dato.tipo == "Socio") {
-            swal(
-                'USUARIO VÁLIDO!',
-                'Usted está registrado en la base de datos!',
-                'success', { button: 'aceptar', }
-            ).then(function() {
+            swal({
+                title: 'USUARIO VÁLIDO',
+                text: 'Usted está registrado/a',
+                timer: 1000,
+                showConfirmButton: false                   
+             }).then(function() {
                 console.log(dato);
                 localStorage.setItem("hora", dato.hora);
                 localStorage.setItem("usuario", dato.Usuario);
@@ -39,11 +41,12 @@ function EnviarDatos() {
             localStorage.setItem("usuario", dato.Usuario);
             localStorage.setItem("rol", dato.tipo);
             localStorage.setItem("token", dato.token);
-            swal(
-                'USUARIO VÁLIDO!',
-                'Usted está registrado en la base de datos!',
-                'success'
-            ).then(function() {
+            swal({
+                title: 'USUARIO VÁLIDO',
+                text: 'Usted está registrado/a',
+                timer: 2000,
+                showConfirmButton: false                   
+             }).then(function() {
                 window.location.replace("../enlaces/restauranteMozo.html");
             }, function() {
                 swal('Ocurrio algo inesperado!');
@@ -53,11 +56,12 @@ function EnviarDatos() {
             localStorage.setItem("usuario", dato.Usuario);
             localStorage.setItem("rol", dato.tipo);
             localStorage.setItem("token", dato.token);
-            swal(
-                'USUARIO VÁLIDO!',
-                'Usted está registrado en la base de datos!',
-                'success'
-            ).then(function() {
+            swal({
+                title: 'USUARIO VÁLIDO',
+                text: 'Usted está registrado/a',
+                timer: 2000,
+                showConfirmButton: false                   
+             }).then(function() {
                 window.location.replace("../enlaces/restauranteCocinero.html");
             }, function() {
                 swal('Ocurrio algo inesperado!');
@@ -67,11 +71,12 @@ function EnviarDatos() {
             localStorage.setItem("usuario", dato.Usuario);
             localStorage.setItem("rol", dato.tipo);
             localStorage.setItem("token", dato.token);
-            swal(
-                'USUARIO VÁLIDO!',
-                'Usted está registrado en la base de datos!',
-                'success'
-            ).then(function() {
+            swal({
+                title: 'USUARIO VÁLIDO',
+                text: 'Usted está registrado/a',
+                timer: 2000,
+                showConfirmButton: false                   
+             }).then(function() {
                 window.location.replace("../enlaces/restauranteBartender.html");
             }, function() {
                 swal('Ocurrio algo inesperado!');
@@ -81,11 +86,12 @@ function EnviarDatos() {
             localStorage.setItem("usuario", dato.Usuario);
             localStorage.setItem("rol", dato.tipo);
             localStorage.setItem("token", dato.token);
-            swal(
-                'USUARIO VÁLIDO!',
-                'Usted está registrado en la base de datos!',
-                'success'
-            ).then(function() {
+            swal({
+                title: 'USUARIO VÁLIDO',
+                text: 'Usted está registrado/a',
+                timer: 2000,
+                showConfirmButton: false                   
+             }).then(function() {
                 window.location.replace("../enlaces/restauranteCervecero.html");
             }, function() {
                 swal('Ocurrio algo inesperado!');
